Rebind <up>/<down> to perform searching.
By default up/down are bound to previous-history
and next-history respectively. The following does the
same but gives the extra functionality where if you
type any text (or more accurately, if there is any text
between the start of the line and the cursor),
the subset of the history starting with that text
is searched (like 4dos for e.g.).
Note to get rid of a line just Ctrl-C