Use LSP for go-mode.
If `lsp-mode` is available then configure it for *.go files, allowing
jumping to definitions, docs on UI, and formatting on save.

If `lsp-mode` is not available then we lose the jump-to-definition
(bound to `M-SPC`) which we previous had, but things should still
work and `go-mode` still be used for `*.go` files.