Allow linkifier to use capture groups.
I realized that it would be nice to write this:

* YOUTUBE-RyTlSAkoGJg

This would work with the following config:

("\\\<YOUTUBE-[0-9a-zA-Z]+\\\>" "https://www.youtube.com/watch?v=%s")

However the end result is that it would send you to a broken page,
because the full text of the button/capture would be added to the link:

https://www.youtube.com/watch?v=YOUTUBE-RyTlSAkoGJg

To avoid that we now support the use of _one_ capture-group:

("\\\<YOUTUBE-\\([0-9a-zA-Z]+\\)\\\>" "https://www.youtube.com/watch?v=%s")

When present the first capture group \\\(...\\\) is used instead of
the literal match.  This allows the youtube example to work as expected.